Added a DataNode interface, Added Reading (#16)
* Added DataNode interface

This interface will be used to create objects to write to the database. This should force proper labeling of datapoints to keep our data organized

* Refactoring, Config

Refactored Data Classes into appropriate package.

Added configuration for tests to write to a seperate branch of the
database.

* Added reading

Implemented a "latch"-based general reading logic through TableReader. A
DataNode object can be  used to "latch" onto the database, meaning it's
value will be kept updated as values in the database change. This
approach is used to utilize Firebase's asynchronous nature.

* Minor: added exception handling test

Added a unit test to assert that an appropriate exception is thrown by
TableReader if an invalid label is specified

* Custom logic insertion when reading

Added a function to DummyNode that will be called whenever the data it
is latched onto is changed - that way, we can add custom logic to
execute when that happens.

There are examples of how this is used in the unit tests (see
TableReaderTest.java).